Added code for Fish Capacity Upgrade and Hook Depth Upgrade
The 2 upgrades available for the game is increasing the amount of fish that the hook can hold, and the depth that the hook can go to. By making variables for the level and condition, I can correlate each level to the amount of fish and the depth that the hook can go to. I also made it so that it is impossible to upgrade levels without enough money.